Clear listener member variable on StartLiveData
Canceling MonitorLiveData effectively relies on
the listener being promptly destructed. The
assumption is that the listener instance is only
referred to by the monitor algorithm.
However, algorithm objects can persist longer
so there is no guarantee that MonitorLiveData will
be the only reference. Reset the StartLiveData
reference after it is handed to MonitorLiveData.